refactor: ScreeningSelectListener 객체 생성으로 역할 분리
